+++ bindings/matrix-sdk-ffi/CHANGELOG.md
@@ -31,6 +31,9 @@ All notable changes to this project will be documented in this file.
 
 ### Features
 
+- Add `Encryption::recover_and_fix_backup` to automatically fix key storage backup if the
+  private backup decryption key is missing, invalid or inconsistent with the public key.
+  ([#6252](https://github.com/matrix-org/matrix-rust-sdk/pull/6252))
 - Add support for [MSC3489](https://github.com/matrix-org/matrix-spec-proposals/pull/3489)  
   live location sharing through a new `TimelineItemContent::LiveLocation` variant.
   ([#6232](https://github.com/matrix-org/matrix-rust-sdk/pull/6232))